feat: support mode with many windows
