+++ dashboards/customized/influxdb_v3/plugin/window.py
@@ -1,0 +1,24 @@
+from dataclasses import dataclass
+from datetime import datetime
+
+from .utils import to_iso_date
+
+
+@dataclass
+class Window:
+    start: datetime
+    end: datetime
+
+    def display(self) -> str:
+        return f"[{self.start_s} .. {self.end_s})"
+
+    def in_window_sql(self) -> str:
+        return f"'{self.start_s}' <= time AND time < '{self.end_s}'"
+
+    @property
+    def start_s(self) -> str:
+        return to_iso_date(self.start)
+
+    @property
+    def end_s(self) -> str:
+        return to_iso_date(self.end)
